fix(items): only suppress activation cards when rule automation is on
suppressesActivationCard promises the rule's activation flow will post
its own output, but ruleEventDispatch gates that flow behind the
Auto-Apply Conditions setting (default off). Suppressing without the
dispatch made the activation a silent no-op. Share the gate via a new
isAutoApplyEnabled util and post the normal card when automation is
off.
