+++ src/documents/item/base.svelte.ts
@@ -1,5 +1,6 @@
 import { createSubscriber } from 'svelte/reactivity';
 import { SYSTEM_ID, systemHookName } from '#system';
+import { isAutoApplyEnabled } from '#utils/isAutoApplyEnabled.js';
 import { DamageRoll } from '../../dice/DamageRoll.js';
 import { ItemActivationManager } from '../../managers/ItemActivationManager.js';
 import { RulesManager } from '../../managers/RulesManager.js';
@@ -166,6 +167,10 @@ class NimbleBaseItem<ItemType extends SystemItemTypes = SystemItemTypes> extends
 	): boolean {
 		if (rolls.length > 0) return false;
 		if ((activation?.effects?.length ?? 0) > 0) return false;
+		// A suppressing rule's replacement output is produced by its activation
+		// flow, which ruleEventDispatch only runs when automation is enabled.
+		// With automation off, the card is the sole record of the activation.
+		if (!isAutoApplyEnabled()) return false;
 		return [...this.rules.values()].some(
 			(rule) => !rule.disabled && rule.suppressesActivationCard(),
 		);
